+++ tests/isfloat.test.js
@@ -4,6 +4,15 @@ test("sends null to isFloat", () => {
   expect(isFloat(null)).toBe(false);
 });
 
+test("sends undefined to isFloat", () => {
+  expect(isFloat(undefined)).toBe(false);
+});
+
+const s1 = Symbol();
+test("sends symbol to isFloat", () => {
+  expect(isFloat(s1)).toBe(false);
+});
+
 test("sends true to isFloat", () => {
   expect(isFloat(true)).toBe(false);
 });