+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.15.0 (January 11th 2021)
+
+- Add isNil() method to test null and undefined values 
+- Add isSymbol() to pimitive checkers 
+
 # 2.14.0 (January 10th 2021)
 
 - Ch.js becomes CheckHard.js